+++ BIOS/driver/vga/vga.hpp
@@ -2,4 +2,4 @@
 #include "../../kernel/types.hpp"
 
 void vga(const char* text, uint8_t color, uint8_t newline);
-void vgaHex(uint64_t value, uint8_t color);
+void vgaHex(uint32_t value, uint8_t color);
